refactor(analysis-types): split fun DTOs
